This replaces uses of the deprecated DataSource::resolve_post_object with deferred loading. The signature of the deprecated function changed in WPGraphQL v1.1.0 to enforce integers instead of just suggesting integers, so feels like the right time to get this updated in this extension. The deprecated methods will probably be removed sooner than later anyway.
I didn't test every field, but on quick testing this works the way I'd expect it to.

You probably know more about some of the nuances of Yoast and the data it's pulling, so you should probably double check that this leads to the behavior you're expecting.

Happy to chat if needed to get this over the finish line.

Thanks!